# replica/table: check memtable before discarding tombstone during read