@lkshminarayanan 2024-10-07
Changelog V2 :
 - rebased on top of latest
 - Updated code to search only the storage group that has the token (Ref : [comment](https://github.com/scylladb/scylladb/pull/20985#discussion_r1790202731))