@lkshminarayanan 2024-10-09
Changelog V4 : Fixes this https://github.com/scylladb/scylladb/pull/20985#discussion_r1792374138