@avikivity 2024-10-09
Could we not move the compacting reader behind the memtable+sstable merge?