@raphaelsc 2024-10-09
> Could we not move the compacting reader behind the memtable+sstable merge?

I thought about that, but we cannot do it because the compacting reader happens in the cache data source, to avoid populating it with garbage. Then the cache reader has later its data combined with memtable.