@lkshminarayanan 2024-10-18
Changelog V5 : 
 - Track maximum seen timestamp rather than actual maximum in compaction_group