@denesb 2024-10-23
> `build/Dev/scylla perf-simple-query --smp=1 --flush` :

In the future, please use `release` for performance measurements.